packages/torch4d/torch4d-0.1.2.tar.gz/torch4d-0.1.2/tests/test_functional.py
""" Testing the torch4d.torch4d Modules """

import pytest
from unittest.mock import patch
from torch4d.functional import max_pool4d, drop_block4d, spatial_dropout4d
import torch

class Tests:
    """ Class defines tests. """
    def test_max_pool4d(self):
        input = torch.randn(32, 400, 2, 8, 8)
        out, _ = max_pool4d(input, 1, 1)
        assert out.shape == torch.Size([32, 400, 2, 8, 8]) 

    def test_drop_block4d(self):
        input = torch.randn(32, 1, 400, 2, 8, 8)
        inputOldShape = input.shape
        drop_block4d(input, p = 0.2, block_size = (100, 1, 2, 2), inplace = True)
        assert input.shape == inputOldShape

    def test_spatial_dropout4d(self):
        input = torch.randn(32, 100, 400, 2, 8, 80) # 100 channels
        OldNoChannels = input.shape[1]
        spatial_dropout4d(input, p=0.2, inplace=True)
        NewNoChannels = input.shape[1]
        assert NewNoChannels <= OldNoChannels